Revert "Preferring to pull out process related classes into their own package. Renaming them to be more generic as well. Nothing about them necessarily requires FFMPEG."
This reverts commit 10ed2d50640f3d7895da3d6654b8bb6e711fcf39.
