+++ jave-core/src/main/java/ws/schild/jave/Encoder.java
@@ -30,9 +30,6 @@ import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import ws.schild.process.ProcessLocator;
-import ws.schild.process.ProcessWrapper;
-
 /**
  * Main class of the package. Instances can encode audio and video streams.
  *
@@ -67,14 +64,14 @@ public class Encoder {
     /**
      * The locator of the ffmpeg executable used by this encoder.
      */
-    private final ProcessLocator locator;
+    private final FFMPEGLocator locator;
     
     /**
      * The executor used to do the conversion
      * Is saved here, so we can abort the conversion process
      * 
      */
-    private ProcessWrapper ffmpeg;
+    private FFMPEGExecutor ffmpeg;
     
     /**
      * List of unhandled messages from ffmpeng run
@@ -90,12 +87,12 @@ public class Encoder {
     }
 
     /**
-     * It builds an encoder with a custom {@link ProcessLocator}.
+     * It builds an encoder with a custom {@link FFMPEGLocator}.
      *
      * @param locator The locator picking up the ffmpeg executable used by the
      * encoder.
      */
-    public Encoder(ProcessLocator locator) {
+    public Encoder(FFMPEGLocator locator) {
         this.locator = locator;
     }
 
@@ -137,7 +134,7 @@ public class Encoder {
      */
     protected String[] getCoders(boolean encoder, boolean audio) throws EncoderException {
         ArrayList<String> res = new ArrayList<>();
-        ProcessWrapper localFFMPEG = locator.createExecutor();
+        FFMPEGExecutor localFFMPEG = locator.createExecutor();
         localFFMPEG.addArgument(encoder ? "-encoders" : "-decoders");
         try
         {
@@ -253,7 +250,7 @@ public class Encoder {
      */
     protected String[] getSupportedCodingFormats(boolean encoding) throws EncoderException {
         ArrayList<String> res = new ArrayList<>();
-        ProcessWrapper localFFMPEG = locator.createExecutor();
+        FFMPEGExecutor localFFMPEG = locator.createExecutor();
         localFFMPEG.addArgument("-formats");
         try
         {
