+++ jave-core/src/main/java/ws/schild/jave/MultimediaObject.java
@@ -10,9 +10,6 @@ import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import ws.schild.process.ProcessLocator;
-import ws.schild.process.ProcessWrapper;
-
 public class MultimediaObject {
 
     /**
@@ -57,7 +54,7 @@ public class MultimediaObject {
     /**
      * The locator of the ffmpeg executable used by this extractor.
      */
-    private final ProcessLocator locator;
+    private final FFMPEGLocator locator;
 
     private File inputFile;
     private URL  inputURL;
@@ -135,13 +132,13 @@ public class MultimediaObject {
     }
     
     /**
-     * It builds an extractor with a custom {@link ProcessLocator}.
+     * It builds an extractor with a custom {@link FFMPEGLocator}.
      *
      * @param input Input file for creating MultimediaObject
      * @param locator The locator picking up the ffmpeg executable used by the
      * extractor.
      */
-    public MultimediaObject(File input, ProcessLocator locator) {
+    public MultimediaObject(File input, FFMPEGLocator locator) {
         this.locator = locator;
         this.inputFile = input;
     }
@@ -160,7 +157,7 @@ public class MultimediaObject {
             EncoderException {
         if (isURL() || inputFile.canRead())
         {
-            ProcessWrapper ffmpeg = locator.createExecutor();
+            FFMPEGExecutor ffmpeg = locator.createExecutor();
             ffmpeg.addArgument("-i");
             if (isURL())
             {
